Rename mongoose to squeasel
A mongoose looks like a combination between a squirrel
and a weasel. This project looks sort of like a Mongoose.

This change renames the files, artifacts, etc, in addition
to renaming all of the APIs from mg_* to sq_*. To make transition
easier, includers may #define MONGOOSE_COMPATIBLE before including
squeasel.h. This uses macros to keep the old APIs working.

Change-Id: I0439a6360b71df52c8bd917e7a9d02344869637a
Reviewed-on: http://gerrit.ent.cloudera.com:8080/434
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Henry Robinson <henry@cloudera.com>