Always dynamic link against SSL
The original source allowed squeasel to load SSL at runtime using
dlopen/dlsym. We always build with SSL support, so it's easier to
clean up this code.

This is also important bceause we'd like to use SSL_CTX_set_option
which requires some constants defined in the OpenSSL headers. With
dlopen, we'd have to hard-code those constants, which seems inadvisable.

Change-Id: I764841b697176adb4e37539e6789aeaeb16620f2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4868
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>