SSL: make SSL library global intiialization optional
This adds a new option 'ssl_global_init' which allows the user to
disable the initialization of the SSL libraries. This is important
in the case that the embedding application is also using OpenSSL and
wants to provide its own locking/thread-ID callbacks. If the application
and Squeasel both try to provide these callbacks, then we may have a
race during initialization or crashes after Squeasel is unloaded.

Change-Id: I5e53a21250350d28ff881adfc9b8dab4e3af58cf