Revert [3/3] "Revert "Always dynamic link against SSL""
This reverts commit b578ce9ce4c3bf4f050035ef762e683054912631.

This goes back to explicitly depending on OpenSSL rather than opening it
using dlopen. To recap history here:

- originally, squeasel used dlopen() to load OpenSSL so that it was an
  optional dependency

- dd40db5c64dc1eed17a50f93407af46269ff8a1e switched to just using it
  explicitly rather than dlopen()ing to simplify the code and allow us
  to use the OpenSSL headers/macros more robustly.

- b578ce9ce4c3bf4f050035ef762e683054912631 reverted back to the dlopen()
  approach because RHEL 6.6 made a breaking change to the OpenSSL ABI.
  Thus, without dynamic linking, a binary built on 6.4 wouldn't work on
  6.6 and vice versa. We wanted to support both operating systems, so
  dlopen() was an appropriate workaround.

Now we're going back to the explicit approach because both Kudu and
Impala (the only known embedders of Squeasel) have inherited their own
explicit dependencies on OpenSSL. Additionally, el6.4 is old enough that
we are less concerned about supporting it at this point.

Change-Id: Ibd53e228195ab82bbfe39a99d88b0d796f9ab958