Don't send HTTP 500 when closing keepalive connection
When a client uses keepalive, the server will time out those connections
after a configurable (30s default) timeout. In this case, the server
gets a timeout error. Prior to this patch, this would result in the
server sending an HTTP 500 error followed by a TCP "FIN" segment.

Some clients wouldn't see the connection was shut down until their next
HTTP request, at which point they would receive the error "500" from a
buffer and assume it was in response to their request, rather than due
to the keepalive shutdown.

Other HTTP servers just send the FIN without any associated response.
This patch changes squeasel to be the same.

This fixes the behavior of Apache Knox when running as a proxy in front
of squeasel -- prior to this patch, if I requested a page, then waited
30 seconds, and requested another, I'd see an error 500. Now, Knox
properly re-fetches a new page.

Change-Id: I2fd4f8e03d611472191a975b59e024e0cd637f50