Fix shutdown to proceed quickly when keepalive is enabled
Prior to this patch, if there was a client connected with keepalive
enabled, sq_stop would block until the configured keepalive timeout.
This would make shutdown quite slow in Kudu tests like webserver-test if
keepalive was enabled.

The fix uses poll() to wait for the either the request or for shutdown.

Change-Id: I7d3a3a1903b7b20a798e1df703d548cf70af1440