Allow request callbacks to specify whether the connection is in a valid state
In certain cases, a request handler may fail to parse a request (eg if
the content-length is larger than it would like to handle, as in Kudu's
WebserverTest.TestPostTooBig). In this case, keepalive shouldn't be
used, and the connection should be closed.

This patch allows handlers to indicate this status as a different return
value.

Change-Id: Ied767c0af2b068424147dc1585fd6eec04b46932